Fix data not displayed correctly on raid device (using int instead of int64)

+++ app/humanize.cpp
@@ -29,7 +29,7 @@ Humanize::~Humanize()
  */
 QString Humanize::size(qint64 size)
 {
-  float num = size;
+  double num = size;
   QStringList list;
   list << "KB" << "MB" << "GB" << "TB";
